requirements.txt
# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.24.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0

# Data Processing & Analysis
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Natural Language Processing
textblob>=0.17.0
nltk>=3.8.0
spacy>=3.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# UI Frameworks
streamlit>=1.28.0
gradio>=4.0.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Development & Testing
pytest>=7.4.0
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# Optional: Advanced ML Libraries
# Uncomment as needed for advanced features
# open-clip-torch>=2.24.0
# controlnet-aux>=0.4.0
# xformers>=0.0.22  # For memory-efficient attention (GPU recommended)
# wandb>=0.16.0     # Experiment tracking
# safetensors>=0.4.0  # Safe model loading

# Optional: Cloud/API Integration
# boto3>=1.34.0     # AWS SDK
# google-cloud-storage>=2.10.0  # Google Cloud Platform
# openai>=1.0.0     # OpenAI API (updated to v1+)
